fastapi/requirements.txt
sphinx==3.5.*
sphinx_rtd_theme

fastapi
pydantic
sqlalchemy
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
